restructured to all xarray, utils to subpackage
