+++ tobac/analysis/centerofgravity.py
@@ -4,6 +4,10 @@
 import logging
 import os
 
+from tobac.utils.convert import xarray_to_iris,iris_to_xarray
+
+
+@xarray_to_iris
 def cog_cell(cell,Tracks=None,M_total=None,M_liquid=None,
              M_frozen=None,
              Mask=None,
@@ -72,7 +76,7 @@ def cog_cell(cell,Tracks=None,M_total=None,M_liquid=None,
 
 
 
-
+@xarray_to_iris
 def calculate_cog(tracks,mass,mask):
     '''Caluclate center of gravity and mass forech tracked cell.
 
@@ -120,7 +124,7 @@ def calculate_cog(tracks,mass,mask):
 
     return tracks_out
     
-
+@xarray_to_iris
 def calculate_cog_untracked(mass,mask):
     '''Caluclate center of gravity and mass for untracked domain parts.
 
@@ -167,7 +171,7 @@ def calculate_cog_untracked(mass,mask):
     
     return tracks_out
 
-
+@xarray_to_iris
 def calculate_cog_domain(mass):
     '''Caluclate center of gravity and mass for entire domain.
 
@@ -209,7 +213,7 @@ def calculate_cog_domain(mass):
 
     return tracks_out
 
-
+@xarray_to_iris
 def center_of_gravity(cube_in):
     '''Caluclate center of gravity and sum of quantity.
