+++ tobac/themes/tobac_v1/__init__.py
@@ -1,6 +1,4 @@
-from .segmentation import segmentation_3D, segmentation_2D,watershedding_3D,watershedding_2D
+from .mask_segmentation import segmentation, segmentation_2D, segmentation_3D
 from .feature_detection import feature_detection_multithreshold
 from .tracking import linking_trackpy
-from .wrapper import maketrack
-from .wrapper import tracking_wrapper
